fixed service user
