+++ src/main/java/ma/m3achaba/plantes/config/JwtAuthFilter.java
@@ -23,9 +23,6 @@ import java.io.IOException;
 public class JwtAuthFilter extends OncePerRequestFilter {
 
     private final JwtUtils jwtService;
-
-    @Autowired
-    @Lazy  // Ajoutez @Lazy ici pour éviter la dépendance circulaire
     private final UserService accountService;
 
     @Override
